requirement.txt
numpy
pandas
python
torch
transformer
sklearn

